Add files via upload
An Outline of what each Java file does:
AudioMixer -> An old version of Source Audio, made to find and connect Audio Drivers and Lines.
BeatMachine -> A secondary keyboard for inputting any sound into the gridpads.
Client -> The code to create a member to join a chat. Connects a Socket to the ServerSocket.
ClientHandler -> Handles all clients, and the messages being relayed to the Server and back to the clients.
Forum -> the UI and listeners for the Forum tab. This is a panel that is replaced into the Center panel of the Frame on the Template page.
MainFeed -> The UI and listeners for the Feed tab. This is a panel that is replaced into the center panel of the frame of the template page.
mainUI -> Old version UI, not used.
MidiInstrument -> This is the UI, listeners,  and audio/midi connections for the midi keyboard that is used.
Pad -> The Jpanel for the BeatMachine buttons.
Profile -> The UI, Listeners, and database data for the user.
Search ->  The UI, Listeners,  for the Search page for Users, chats Forums.
Server -> The sever creation class.
SourceAudio ->  Made to find and connect Audio Drivers and Lines.
Template -> The main used window for the creation of the template. The Chat window is create within this file.
TextMessage -> The Ui for the text boxes, as well as its data included. Listeners, and the text from the user are taken into this class.